products/qr/requirements.txt
streamlit
qrcode
Pillow
pydrive